requirements.txt
Django==2.2
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.2
Markdown==3.3.3
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
tensorflow==2.3.1
sentry-sdk==0.20.3
pillow==8.1.0
